requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
billiard==3.6.4.0
boto3==1.26.142
botocore==1.29.142
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.2
dataclasses-json==0.5.7
Django==4.2.1
django-appconf==1.0.5
django-celery-results==2.5.1
django-cryptography==1.1
django-elasticsearch-dsl==7.3
django-elasticsearch-dsl-drf==0.22.5
django-filter==23.2
django-nine==0.2.7
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
elasticsearch==7.17.9
elasticsearch-dsl==7.4.1
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
jmespath==1.0.1
kombu==5.2.4
langchain==0.0.146
Markdown==3.4.3
markdown-it-py==2.2.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.3
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
rich==13.3.5
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==1.4.48
sqlparse==0.4.4
stripe==5.4.0
tenacity==8.2.2
tqdm==4.65.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
